# backend/src/routes/kimi-cluster.ts
/**
 * kimi-cluster.ts — KIMI 集群编排路由
 * 暴露 KimiClusterOrchestrator 核心功能
 */

import { Router } from 'express';
import { kimiCluster } from '../KimiClusterOrchestrator';

const router = Router();

// GET /api/kimi-cluster/status — 集群状态
router.get('/status', async (req, res) => {
  try {
    const reply = await kimiCluster.handleRpc({
      id: `req-${Date.now()}`,
      x: { type: 'kimi-cluster', action: 'getStatus' },
      y: { service: 'kimi-cluster', version: '1.0' },
      z: { priority: 1 },
      semantic_payload: { action: 'getStatus' },
      protocol_adapter: 'REST',
    } as any);

    res.json({
      success: true,
      data: reply.semantic_payload,
      timestamp: new Date().toISOString(),
    });
  } catch (err: any) {
    res.status(500).json({ success: false, error: err.message });
  }
});

// GET /api/kimi-cluster/patterns — 活动模式分析
router.get('/patterns', async (req, res) => {
  try {
    const agentId = req.query.agentId as string | undefined;
    const reply = await kimiCluster.handleRpc({
      id: `req-${Date.now()}`,
      x: { type: 'kimi-cluster', action: 'getPatterns' },
      y: { service: 'kimi-cluster', version: '1.0' },
      z: { priority: 1 },
      semantic_payload: { action: 'getPatterns', agentId },
      protocol_adapter: 'REST',
    } as any);

    res.json({
      success: true,
      data: reply.semantic_payload,
      timestamp: new Date().toISOString(),
    });
  } catch (err: any) {
    res.status(500).json({ success: false, error: err.message });
  }
});

// POST /api/kimi-cluster/load-balance — 负载均衡决策
router.post('/load-balance', async (req, res) => {
  try {
    const { agentId, taskType, payload } = req.body;

    const reply = await kimiCluster.handleRpc({
      id: `req-${Date.now()}`,
      x: { type: 'kimi-cluster', action: 'getDecision' },
      y: { service: 'kimi-cluster', version: '1.0' },
      z: { priority: 1 },
      semantic_payload: {
        action: 'getDecision',
        agentId: agentId || 'default',
        taskType: taskType || 'chat',
        payload: payload || {},
      },
      protocol_adapter: 'REST',
    } as any);

    res.json({
      success: true,
      data: reply.semantic_payload,
      timestamp: new Date().toISOString(),
    });
  } catch (err: any) {
    res.status(500).json({ success: false, error: err.message });
  }
});

// POST /api/kimi-cluster/endpoints — 添加端点
router.post('/endpoints', async (req, res) => {
  try {
    const { id, baseUrl, apiKey, model, weight, capabilities } = req.body;

    const reply = await kimiCluster.handleRpc({
      id: `req-${Date.now()}`,
      x: { type: 'kimi-cluster', action: 'addEndpoint' },
      y: { service: 'kimi-cluster', version: '1.0' },
      z: { priority: 1 },
      semantic_payload: {
        action: 'addEndpoint',
        endpoint: { id, baseUrl, apiKey, model, weight: weight || 1, capabilities: capabilities || [] },
      },
      protocol_adapter: 'REST',
    } as any);

    res.json({
      success: true,
      data: reply.semantic_payload,
      timestamp: new Date().toISOString(),
    });
  } catch (err: any) {
    res.status(500).json({ success: false, error: err.message });
  }
});

// DELETE /api/kimi-cluster/endpoints/:id — 移除端点
router.delete('/endpoints/:id', async (req, res) => {
  try {
    const { id } = req.params;

    const reply = await kimiCluster.handleRpc({
      id: `req-${Date.now()}`,
      x: { type: 'kimi-cluster', action: 'removeEndpoint' },
      y: { service: 'kimi-cluster', version: '1.0' },
      z: { priority: 1 },
      semantic_payload: { action: 'removeEndpoint', endpointId: id },
      protocol_adapter: 'REST',
    } as any);

    res.json({
      success: true,
      data: reply.semantic_payload,
      timestamp: new Date().toISOString(),
    });
  } catch (err: any) {
    res.status(500).json({ success: false, error: err.message });
  }
});

export default router;