# openapi.yaml
openapi: 3.0.3
info:
  title: 千界花园 (Thousand Realms Garden) API
  description: '群智协同操作系统 REST API — XYZ 三轴自由连接、多 Agent 编排、统一 LLM 适配


    📊 自动扩展: 145 个新端点已从后端路由同步'
  version: 1.0.0
  contact:
    name: Thousand Realms Garden Team
  license:
    name: MIT
servers:
- url: http://localhost:3001/api
  description: 本地开发环境
- url: https://thousand-realms-garden.onrender.com/api
  description: 生产环境 (Render)
...
paths:
...
  /knowledge:
    get:
      tags:
      - Knowledge
      summary: 获取知识库列表
      security:
      - BearerAuth: []
      responses:
        200:
          description: 知识库列表
    post:
      tags:
      - Knowledge
      summary: 创建知识库
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - name
              properties:
                name:
                  type: string
                description:
                  type: string
                embeddingModel:
                  type: string
                  default: openai/text-embedding-3-small
      responses:
        201:
          description: 创建成功
  /knowledge/{id}/upload:
    post:
      tags:
      - Knowledge
      summary: 上传文档到知识库
      security:
      - BearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
              - file
              properties:
                file:
                  type: string
                  format: binary
                  description: 支持 pdf/md/txt/docx
      responses:
        201:
          description: 上传成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      fileId:
                        type: string
                      chunks:
                        type: integer
                      status:
                        type: string
  /knowledge/{id}/search:
    post:
      tags:
      - Knowledge
      summary: 语义检索
      security:
      - BearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - query
              properties:
                query:
                  type: string
                  description: 检索查询
                topK:
                  type: integer
                  default: 5
                threshold:
                  type: number
                  default: 0.7
                  description: 相似度阈值
      responses:
        200:
          description: 检索结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        chunk:
                          type: string
                        source:
                          type: string
                        relevance:
                          type: number
                        page:
                          type: integer
...
  /kimi-cluster/endpoints:
    post:
      tags:
      - Kimi Cluster
      summary: POST /kimi-cluster/endpoints
      operationId: post_kimi-cluster_endpoints
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
  /kimi-cluster/endpoints/{id}:
    delete:
      tags:
      - Kimi Cluster
      summary: DELETE /kimi-cluster/endpoints/{id}
      operationId: delete_kimi-cluster_endpoints_id
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /kimi-cluster/load-balance:
    post:
      tags:
      - Kimi Cluster
      summary: POST /kimi-cluster/load-balance
      operationId: post_kimi-cluster_load-balance
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
  /kimi-cluster/patterns:
    get:
      tags:
      - Kimi Cluster
      summary: GET /kimi-cluster/patterns
      operationId: get_kimi-cluster_patterns
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
  /kimi-cluster/status:
    get:
      tags:
      - Kimi Cluster
      summary: GET /kimi-cluster/status
      operationId: get_kimi-cluster_status
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
  /knowledge/{id}:
    get:
      tags:
      - Knowledge
      summary: GET /knowledge/{id}
      operationId: get_knowledge_id
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    put:
      tags:
      - Knowledge
      summary: PUT /knowledge/{id}
      operationId: put_knowledge_id
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    delete:
      tags:
      - Knowledge
      summary: DELETE /knowledge/{id}
      operationId: delete_knowledge_id
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /knowledge/{id}/documents:
    post:
      tags:
      - Knowledge
      summary: POST /knowledge/{id}/documents
      operationId: post_knowledge_id_documents
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /knowledge/{id}/documents/:docId:
    delete:
      tags:
      - Knowledge
      summary: DELETE /knowledge/{id}/documents/:docId
      operationId: delete_knowledge_id_documents_:docId
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数错误
        '401':
          description: 未认证
        '500':
          description: 服务器错误
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 通过 /auth/login 获取的 JWT Token